refactor(test): move tests to live alongside file
