+++ src/routes/admin/profiles/editProfile.test.ts
@@ -1,11 +1,11 @@
-import { AMTConfiguration } from '../models/Rcs'
+import { AMTConfiguration } from '../../../models/Rcs'
 import {
   handleAMTPassword,
   handleMEBxPassword,
   handleGenerateRandomPassword,
   handleGenerateRandomMEBxPassword,
   getUpdatedData
-} from '../routes/admin/profiles/editProfile'
+} from './editProfile'
 
 test('test handleAMTpassword when the request body amtPassword is null or undefined', () => {
   const newConfig: AMTConfiguration = { profileName: 'acm', activation: 'acmactivate', tenantId: '' }
