+++ src/utils/maintenance/synchronizeTime.test.ts
@@ -1,13 +1,13 @@
-import { ClientManager } from '../ClientManager'
-import Logger from '../Logger'
-import { NodeForge } from '../NodeForge'
-import { ClientResponseMsg } from '../utils/ClientResponseMsg'
-import { RPSError } from '../utils/RPSError'
-import { WSManProcessor } from '../WSManProcessor'
+import { ClientManager } from '../../ClientManager'
+import Logger from '../../Logger'
+import { NodeForge } from '../../NodeForge'
+import { ClientResponseMsg } from '../ClientResponseMsg'
+import { RPSError } from '../RPSError'
+import { WSManProcessor } from '../../WSManProcessor'
 import { v4 as uuid } from 'uuid'
-import { synchronizeTime } from '../utils/maintenance/synchronizeTime'
-import { ClientObject } from '../models/RCS.Config'
-import { AMTUserName } from '../utils/constants'
+import { synchronizeTime } from './synchronizeTime'
+import { ClientObject } from '../../models/RCS.Config'
+import { AMTUserName } from '../constants'
 
 const clientManager = ClientManager.getInstance(new Logger('ClientManager'))
 const nodeForge = new NodeForge()
