+++ src/utils/maintenance/updateAMTAdminPassword.test.ts
@@ -1,16 +1,16 @@
-import { ClientManager } from '../ClientManager'
-import Logger from '../Logger'
-import { NodeForge } from '../NodeForge'
-import { ClientResponseMsg } from '../utils/ClientResponseMsg'
-import { RPSError } from '../utils/RPSError'
-import { WSManProcessor } from '../WSManProcessor'
+import { ClientManager } from '../../ClientManager'
+import Logger from '../../Logger'
+import { NodeForge } from '../../NodeForge'
+import { ClientResponseMsg } from '../ClientResponseMsg'
+import { RPSError } from '../RPSError'
+import { WSManProcessor } from '../../WSManProcessor'
 import { v4 as uuid } from 'uuid'
-import { updateAMTAdminPassword } from '../utils/maintenance/updateAMTAdminPassword'
-import { ClientObject } from '../models/RCS.Config'
-import { Configurator } from '../Configurator'
-import { Validator } from '../Validator'
-import { config } from './helper/Config'
-import { EnvReader } from '../utils/EnvReader'
+import { updateAMTAdminPassword } from './updateAMTAdminPassword'
+import { ClientObject } from '../../models/RCS.Config'
+import { Configurator } from '../../Configurator'
+import { Validator } from '../../Validator'
+import { config } from '../../test/helper/Config'
+import { EnvReader } from '../EnvReader'
 
 EnvReader.GlobalEnvConfig = config
 const clientManager = ClientManager.getInstance(new Logger('ClientManager'))
